// shared/metrics/src/domain/metrics-exporter.interface.ts
export enum ExportFormat {
  /**
   * Prometheus text-based exposition format
   * https://prometheus.io/docs/instrumenting/exposition_formats/
   */
  PROMETHEUS = 'prometheus',

  /**
   * JSON format for HTTP APIs
   */
  JSON = 'json',

  /**
   * Grafana dashboard JSON
   */
  GRAFANA_DASHBOARD = 'grafana_dashboard',

  /**
   * OpenTelemetry format
   */
  OPENTELEMETRY = 'opentelemetry',
}